+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+build 7
+-------
++ added configuration
++ shortened standard long-press duration to 0.8s
++ uninstaller now completely removes the hack
+
 build 6
 -------
 + moved to hard-float ABI, neccessary for fw 2.6.1